Making stereo_view into a component.
